EdkRepo: Resolve circular dependency.
Move clone_single_repository() to common_repo_functions.py
to resolve circular dependency.
